+++ sources/loaders/apple2_do_loader/apple2_do_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int Apple2_do_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int Apple2_do_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
