+++ sources/loaders/cpcdsk_loader/cpcdsk_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int CPCDSK_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int CPCDSK_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
