+++ sources/loaders/d64_loader/d64_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int D64_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int D64_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
