+++ sources/loaders/dms_loader/dms_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int DMS_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int DMS_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
