+++ sources/loaders/emuii_loader/emuii_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int EMUII_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int EMUII_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
