+++ sources/loaders/fd_loader/fd_loader.h
@@ -25,4 +25,4 @@
 //
 */
 
-int FD_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,unsigned long infotype,void * returnvalue);
+int FD_libGetPluginInfo(HXCFE_IMGLDR * imgldr_ctx,uint32_t infotype,void * returnvalue);
